detect: move file hash common code
